# perf: optimize CI runtime from 7m20s to 1s for PRs