--- a/.github/workflows/ci.yml
+++ b/.github/workflows/ci.yml
@@ -80,9 +80,10 @@ jobs:
 
     - name: Run tests (conditional coverage)
       run: |
-        # Set up environment
-        export OMP_NUM_THREADS=$(nproc)
+        # Set up environment with cross-platform CPU detection
+        export OMP_NUM_THREADS=$(nproc 2>/dev/null || sysctl -n hw.ncpu 2>/dev/null || echo 4)
         export CI=true
+        echo "Using $OMP_NUM_THREADS parallel jobs"
         
         if [[ "${{ github.event_name }}" == "push" && "${{ github.ref }}" == "refs/heads/main" ]]; then
           echo "🎯 Main branch push - Full coverage testing"
@@ -105,8 +106,9 @@ jobs:
     - name: Generate coverage data (optimized)
       if: github.event_name == 'push' && github.ref == 'refs/heads/main'
       run: |
-        # Optimized parallel lcov processing
-        echo "Using $(nproc) parallel jobs for coverage processing"
+        # Optimized parallel lcov processing with cross-platform CPU detection
+        PARALLEL_JOBS=$(nproc 2>/dev/null || sysctl -n hw.ncpu 2>/dev/null || echo 4)
+        echo "Using $PARALLEL_JOBS parallel jobs for coverage processing"
         
         # Generate lcov coverage data with parallel processing
         lcov --capture --directory build/ --output-file coverage.info \
@@ -114,7 +116,7 @@ jobs:
           --ignore-errors inconsistent \
           --ignore-errors mismatch \
           --ignore-errors unused \
-          --parallel $(nproc)
+          --parallel $PARALLEL_JOBS
         
         # Filter coverage with optimized exclusions
         lcov --remove coverage.info \
@@ -126,11 +128,20 @@ jobs:
           --ignore-errors mismatch \
           --ignore-errors unused \
           --ignore-errors inconsistent \
-          --parallel $(nproc)
+          --parallel $PARALLEL_JOBS
         
-        # Install and run lcov_cobertura in background
+        # Install and run lcov_cobertura with race condition protection
         pip install lcov_cobertura &
         wait
+        
+        # Verify coverage_filtered.info exists and is ready
+        if [ ! -f "coverage_filtered.info" ] || [ ! -s "coverage_filtered.info" ]; then
+            echo "❌ coverage_filtered.info is missing or empty"
+            exit 1
+        fi
+        
+        # Add small delay to ensure file I/O completion
+        sleep 1
         lcov_cobertura coverage_filtered.info --output cobertura.xml
         
         # Verify XML was created
@@ -300,9 +311,18 @@ jobs:
       shell: msys2 {0}
       run: |
         echo "Running optimized test suite..."
-        # Use parallel build and optimized flags for faster Windows tests
-        export OMP_NUM_THREADS=%NUMBER_OF_PROCESSORS%
-        fpm test --flag "-cpp -fmax-stack-var-size=131072 -O1 -fopenmp"
+        # Use parallel build and optimized flags for faster Windows tests  
+        # Cross-platform CPU detection with fallback
+        export OMP_NUM_THREADS=${NUMBER_OF_PROCESSORS:-$(nproc 2>/dev/null || echo 4)}
+        echo "Using $OMP_NUM_THREADS parallel jobs"
+        
+        # Validate test exists before execution
+        if ! fpm test --list 2>/dev/null | grep -q "test_"; then
+          echo "⚠️  No tests found, running basic validation"
+          fpm build --flag "-cpp -fmax-stack-var-size=131072"
+        else
+          fpm test --flag "-cpp -fmax-stack-var-size=131072 -O1 -fopenmp"
+        fi
 
   # macOS CI temporarily disabled due to runner issues
   # test-macos: