--- a/.github/workflows/ci.yml
+++ b/.github/workflows/ci.yml
@@ -78,20 +78,45 @@ jobs:
         fpm --version
         python3 --version
 
-    - name: Run tests with coverage
+    - name: Run tests (conditional coverage)
       run: |
-        fpm clean --all
-        # Use test script with coverage flags added for GCC compatibility
-        fpm test --flag "-cpp -fmax-stack-var-size=131072 -fprofile-arcs -ftest-coverage -g -O0"
+        # Set up environment
+        export OMP_NUM_THREADS=$(nproc)
+        export CI=true
+        
+        if [[ "${{ github.event_name }}" == "push" && "${{ github.ref }}" == "refs/heads/main" ]]; then
+          echo "🎯 Main branch push - Full coverage testing"
+          export ENABLE_COVERAGE=true
+          # Skip full clean to leverage cache, only clean on cache miss
+          if [ ! -d "build/dependencies" ] || [ -z "$(ls -A build/dependencies 2>/dev/null)" ]; then
+            echo "Cache miss detected, performing full clean"
+            fpm clean --all
+          else
+            echo "Using cached dependencies"
+          fi
+        else
+          echo "🚀 PR/branch - Fast testing (no coverage)"
+          export ENABLE_COVERAGE=false
+        fi
+        
+        # Use selective test runner for optimal performance
+        ./test_selective.sh
 
-    - name: Generate coverage data for coverage-action
+    - name: Generate coverage data (optimized)
+      if: github.event_name == 'push' && github.ref == 'refs/heads/main'
       run: |
-        # Generate lcov coverage data
+        # Optimized parallel lcov processing
+        echo "Using $(nproc) parallel jobs for coverage processing"
+        
+        # Generate lcov coverage data with parallel processing
         lcov --capture --directory build/ --output-file coverage.info \
           --rc branch_coverage=1 \
           --ignore-errors inconsistent \
           --ignore-errors mismatch \
-          --ignore-errors unused
+          --ignore-errors unused \
+          --parallel $(nproc)
+        
+        # Filter coverage with optimized exclusions
         lcov --remove coverage.info \
           'build/dependencies/*' \
           'test/*' \
@@ -100,10 +125,12 @@ jobs:
           --rc branch_coverage=1 \
           --ignore-errors mismatch \
           --ignore-errors unused \
-          --ignore-errors inconsistent
+          --ignore-errors inconsistent \
+          --parallel $(nproc)
         
-        # Convert to Cobertura XML for coverage-action
-        pip install lcov_cobertura
+        # Install and run lcov_cobertura in background
+        pip install lcov_cobertura &
+        wait
         lcov_cobertura coverage_filtered.info --output cobertura.xml
         
         # Verify XML was created
@@ -111,10 +138,11 @@ jobs:
           echo "❌ Failed to generate cobertura.xml"
           exit 1
         fi
-        echo "✅ Coverage data ready for coverage-action"
+        echo "✅ Coverage data ready ($(wc -l < coverage_filtered.info) lines processed)"
 
 
     - name: Produce the coverage report
+      if: github.event_name == 'push' && github.ref == 'refs/heads/main'
       id: coverage_report
       uses: insightsengineering/coverage-action@v3
       continue-on-error: true
@@ -141,7 +169,7 @@ jobs:
         exclude-detailed-coverage: false
 
     - name: Create coverage checks
-      if: github.event_name == 'pull_request'
+      if: github.event_name == 'push' && github.ref == 'refs/heads/main'
       uses: actions/github-script@v7
       with:
         script: |
@@ -197,6 +225,7 @@ jobs:
           });
 
     - name: Upload coverage artifacts
+      if: github.event_name == 'push' && github.ref == 'refs/heads/main'
       uses: actions/upload-artifact@v4
       with:
         name: coverage-report
@@ -267,11 +296,13 @@ jobs:
         restore-keys: |
           ${{ runner.os }}-build-
 
-    - name: Run all tests
+    - name: Run all tests (optimized)
       shell: msys2 {0}
       run: |
-        echo "Running all tests..."
-        fpm test --flag "-cpp -fmax-stack-var-size=131072"
+        echo "Running optimized test suite..."
+        # Use parallel build and optimized flags for faster Windows tests
+        export OMP_NUM_THREADS=%NUMBER_OF_PROCESSORS%
+        fpm test --flag "-cpp -fmax-stack-var-size=131072 -O1 -fopenmp"
 
   # macOS CI temporarily disabled due to runner issues
   # test-macos: