--- a/test_selective.sh
+++ b/test_selective.sh
@@ -33,8 +33,8 @@ case $TEST_MODE in
         ;;
 esac
 
-# Set parallel compilation
-export OMP_NUM_THREADS=${OMP_NUM_THREADS:-$(nproc)}
+# Set parallel compilation with cross-platform CPU detection
+export OMP_NUM_THREADS=${OMP_NUM_THREADS:-$(nproc 2>/dev/null || sysctl -n hw.ncpu 2>/dev/null || echo 4)}
 echo "Using $OMP_NUM_THREADS parallel jobs"
 
 # Smart caching
@@ -52,9 +52,28 @@ start_time=$(date +%s)
 if [[ "$TEST_MODE" == "fast" ]]; then
     # For PRs, run core tests only to save time
     echo "Running essential test subset for PR validation"
-    fpm test test_frontend_lexer_api test_semantic_simple test_codegen_core_direct --flag "$FLAGS"
+    
+    # Validate each test exists before execution  
+    FAST_TESTS="test_frontend_lexer_api test_semantic_simple test_codegen_core_direct"
+    EXISTING_TESTS=""
+    
+    for test in $FAST_TESTS; do
+        if fpm test --list 2>/dev/null | grep -q "$test"; then
+            EXISTING_TESTS="$EXISTING_TESTS $test"
+        else
+            echo "⚠️  Test $test not found, skipping"
+        fi
+    done
+    
+    if [ -n "$EXISTING_TESTS" ]; then
+        fpm test $EXISTING_TESTS --flag "$FLAGS"
+    else
+        echo "⚠️  No fast tests found, running full test suite"
+        fpm test --flag "$FLAGS"
+    fi
 else
     # Full test suite
+    echo "Running full test suite"
     fpm test --flag "$FLAGS"
 fi